Store categories to a separated collection.

+++ functions/src/index.ts
@@ -9,74 +9,39 @@ admin.initializeApp();
 
 const presetCategories = [
   {
-    id: 'work',
     label: 'Work',
     color: '#ef5350',
+    activities: [{ label: 'General work' }],
   },
   {
-    id: 'learning',
     label: 'Learning',
     color: '#29b6f6',
+    activities: [{ label: 'General learning' }],
   },
   {
-    id: 'exercise',
     label: 'Exercise',
     color: '#9ccc65',
+    activities: [{ label: 'Walking' }, { label: 'Running' }],
   },
   {
-    id: 'entertainment',
     label: 'Entertainment',
     color: '#ffa726',
+    activities: [{ label: 'Video game' }],
   },
   {
-    id: 'rest',
     label: 'Rest',
     color: '#7e57c2',
+    activities: [{ label: 'Sleep' }],
   },
   {
-    id: 'chore',
     label: 'Chore',
     color: '#8d6e63',
+    activities: [{ label: 'House cleaning' }, { label: 'Errands' }],
   },
   {
-    id: 'others',
     label: 'Others',
     color: '#bdbdbd',
-  },
-];
-
-const presetUserData = {
-  categories: presetCategories,
-};
-
-const presetActivities = [
-  {
-    cid: 'work',
-    label: 'General work',
-  },
-  {
-    cid: 'learning',
-    label: 'General learning',
-  },
-  {
-    cid: 'exercise',
-    label: 'Walking',
-  },
-  {
-    cid: 'entertainment',
-    label: 'Video game',
-  },
-  {
-    cid: 'rest',
-    label: 'Sleep',
-  },
-  {
-    cid: 'chore',
-    label: 'Room cleaning',
-  },
-  {
-    cid: 'chore',
-    label: 'Errands',
+    activities: [{ label: 'Any other activity' }],
   },
 ];
 
@@ -88,17 +53,33 @@ async function onCreateUser(user: UserRecord) {
         ' name:' +
         user.displayName
     );
+    const userDoc = {
+      displayName: user.displayName || '',
+      email: user.email || '',
+    };
     const userDocRef = admin.firestore().collection('users').doc(user.uid);
-    await userDocRef.create(presetUserData);
+    await userDocRef.create(userDoc);
 
-    for (const activity of presetActivities) {
-      const docData = {
+    const categoryCollection = admin.firestore().collection('categories');
+    const activityCollection = admin.firestore().collection('activities');
+    for (const category of presetCategories) {
+      const categoryDoc = {
         uid: user.uid,
-        label: activity.label,
-        cid: activity.cid,
-        updated: admin.firestore.FieldValue.serverTimestamp(),
+        label: category.label,
+        color: category.color,
       };
-      await admin.firestore().collection('activities').add(docData);
+      const newCategoryRef = categoryCollection.doc();
+      await newCategoryRef.set(categoryDoc);
+
+      for (const activity of category.activities) {
+        const activityDoc = {
+          uid: user.uid,
+          label: activity.label,
+          cid: newCategoryRef.id,
+          updated: admin.firestore.FieldValue.serverTimestamp(),
+        };
+        await activityCollection.add(activityDoc);
+      }
     }
     functions.logger.log('Successfully added presert for ' + user.uid);
   } catch (e) {
@@ -116,6 +97,15 @@ async function onDeleteUser(user: UserRecord) {
     );
     await admin.firestore().collection('users').doc(user.uid).delete();
 
+    const records = await admin
+      .firestore()
+      .collection('records')
+      .where('uid', '==', user.uid)
+      .get();
+    for (const doc of records.docs) {
+      await doc.ref.delete();
+    }
+
     const activities = await admin
       .firestore()
       .collection('activities')
@@ -125,14 +115,15 @@ async function onDeleteUser(user: UserRecord) {
       await doc.ref.delete();
     }
 
-    const records = await admin
+    const categories = await admin
       .firestore()
-      .collection('records')
+      .collection('categories')
       .where('uid', '==', user.uid)
       .get();
-    for (const doc of records.docs) {
+    for (const doc of categories.docs) {
       await doc.ref.delete();
     }
+
     functions.logger.log('Successfully deleted data for ' + user.uid);
   } catch (e) {
     functions.logger.error('Failed to delete data for uid: ' + user.uid);
